Struct ic_cycles_account_manager::CyclesAccountManager

pub struct CyclesAccountManager { /* fields omitted */ }

Handles any operation related to cycles accounting, such as charging (due to using system resources) or refunding unused cycles.

Implementations

impl CyclesAccountManager

pub fn new(
max_num_instructions: NumInstructions,
max_cycles_per_canister: Option<Cycles>,
own_subnet_type: SubnetType,
subnet_id: SubnetId,
config: CyclesAccountManagerConfig
) -> Self

pub fn subnet_type(&self) -> SubnetType

Returns the subnet type of this CyclesAccountManager.

pub fn canister_creation_fee(&self) -> Cycles

Returns the fee to create a canister in Cycles.

pub fn ingress_message_received_fee(&self) -> Cycles

Returns the fee for receiving an ingress message in Cycles.

pub fn ingress_byte_received_fee(&self) -> Cycles

Returns the fee per byte of ingress message received in Cycles.

pub fn xnet_call_performed_fee(&self) -> Cycles

Returns the fee for performing a xnet call in Cycles.

pub fn xnet_call_bytes_transmitted_fee(&self, payload_size: NumBytes) -> Cycles

Returns the fee per byte of transmitted xnet call in Cycles.

pub fn withdraw_cycles(
&self,
system_state: &mut SystemState,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation,
cycles: Cycles
) -> Result<(), CyclesAccountError>

Subtracts cycles worth of cycles from the canister's balance.

Errors

Returns a CyclesAccountError::CanisterOutOfCycles if the requested amount is greater than the currently available.

pub fn withdraw_execution_cycles(
&self,
system_state: &mut SystemState,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation,
num_instructions: NumInstructions
) -> Result<(), CyclesAccountError>

Subtracts the corresponding cycles worth of the provided num_instructions from the canister's balance.

Errors

Returns a CyclesAccountError::CanisterOutOfCycles if the requested amount is greater than the currently available.

pub fn refund_execution_cycles(
&self,
system_state: &mut SystemState,
num_instructions: NumInstructions
)

Refunds the corresponding cycles worth of the provided num_instructions to the canister's balance.

pub fn charge_for_compute_allocation(
&self,
system_state: &mut SystemState,
compute_allocation: ComputeAllocation,
duration: Duration
) -> Result<(), CyclesAccountError>

Charges the canister for its compute allocation

Errors

Returns a CyclesAccountError::CanisterOutOfCycles if the requested amount is greater than the currently available.

pub fn ingress_induction_cost(
&self,
ingress: &SignedIngressContent
) -> Result<IngressInductionCost, IngressInductionCostError>

Computes the cost of inducting an ingress message.

Returns a tuple containing:

ID of the canister that should pay for the cost.

The cost of inducting the message.

pub fn charge_for_memory(
&self,
system_state: &mut SystemState,
bytes: NumBytes,
duration: Duration
) -> Result<(), CyclesAccountError>

Subtracts the cycles cost of using a bytes amount of memory.

Note: The following charges for memory taken by the canister. It currently takes into account all the pages in the canister's heap and stable memory (among other things). This will be revised in the future to take into account charging for dirty/read pages by the canister.

Errors

Returns a CyclesAccountError::CanisterOutOfCycles if there's not enough cycles to charge for memory.

pub fn withdraw_request_cycles(
&self,
system_state: &mut SystemState,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation,
request: &Request
) -> Result<(), CyclesAccountError>

When sending a request it's necessary to pay for:

The network cost of sending the request payload, which depends on the size (bytes) of the request.

The max cycles max_num_instructions that would be required to process the Response.

The max network cost of receiving the response, since we don't know yet the exact size the response will have.

The leftover cycles is reimbursed after the Response for this request is received and executed. Only at that point will be known how much cycles receiving and executing the Response costs exactly.

Errors

Returns a CyclesAccountError::CanisterOutOfCycles if there is not enough cycles available to send the Request.

pub fn response_cycles_refund(
&self,
system_state: &mut SystemState,
response: &mut Response
)

Refunds the cycles from the response. In particular, adds leftover cycles from the what was reserved when the corresponding Request was sent earlier.

pub fn cycles_balance_above_storage_reserve(
&self,
system_state: &SystemState,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation
) -> Cycles

Amount of cycles above the level reserved for storage costs

pub fn check_max_cycles_can_add(
&self,
system_state: &SystemState,
cycles_to_add: Cycles
) -> Cycles

Returns the maximum amount of Cycles that can be added to a canister's balance taking into account the max_cycles_per_canister value if present.

pub fn add_cycles(&self, system_state: &mut SystemState, cycles_to_add: Cycles)

Adds cycles worth of cycles to the canister's balance. The cycles balance added in a single go is limited to u64::max_value()

pub fn mint_cycles(
&self,
system_state: &mut SystemState,
amount_to_mint: Cycles
) -> Result<(), CyclesAccountManagerError>

Mints amount_to_mint Cycles.

Errors

Returns a CyclesAccountManagerError::ContractViolation if not on a system subnet.

pub fn charge_canister_for_resource_allocation_and_usage(
&self,
log: &ReplicaLogger,
canister: CanisterState,
duration_between_blocks: Duration
) -> Result<CanisterState, CanisterState>

Charges a canister for its resource allocation and usage for the duration specified. If fees were successfully charged, then returns Ok(CanisterState) else returns Err(CanisterState).